Exposed sceneitems
UGH FINE I'LL DO IT. Originally when sources were created in Rerun they were expected to be automatically visible on screen, but that's no longer the case. Exposing sceneitems allows sources to be created without being attached to a scene and then added to one later. This is becoming more like a general-purpose OBS binding...